+++ CHANGELOG.md
@@ -19,6 +19,8 @@ New features and improvements:
   using ``engine="pyogrio"`` (#2788).
 - Added a ``to_wgs84`` keyword to ``to_json`` allowing automatic re-projecting to follow
   the 2016 GeoJSON specification (#416).
+- Improve error messages when accessing the `geometry` attribute of GeoDataFrame without an active geometry column 
+  related to the default name `"geometry"` being provided in the constructor (#2577)
 
 Deprecations and compatibility notes: